Project2 Finished (Updated different text for alert in func) Learned Gradients, Button, Image, Alert
Add an @State property to store the user’s score, modify it when they get an answer right or wrong, then display it in the alert.
Show the player’s current score in a label directly below the flags.
When someone chooses the wrong flag, tell them their mistake in your alert message - something like “Wrong! That’s the flag of France,” for example.