Day19 Challenge Finished
You need to build an app that handles unit conversions: users will select an input unit and an output unit, then enter a value, and see the output of the conversion.
Which units you choose are down to you, but you could choose one of these:
- Length conversion: users choose meters, kilometers, feet, yards, or miles.
So, if you chose meters for source unit and feet for output unit, then entered 10, you’d see 32.81 as the output.