Day 34 Ended
Go back to the Guess the Flag project and add some animation:

When you tap the correct flag, make it spin around 360 degrees on the Y axis.
Make the other two buttons fade out to 25% opacity.
And if you tap on the wrong flag? Well, that’s down to you – get creative!