Day 47 Finished
At the very least, this means there should be a list of all activities they want to track, plus a form to add new activities – a title and description should be enough.

For a bigger challenge, tapping one of the activities should show a detail screen with the description, how many times they have completed it, plus a button incrementing their completion count.

For an even bigger challenge, use Codable and UserDefaults to load and save all your data.